clone: handle overridden main branch names
When cloning a repository without any branches, Git chooses a default
branch name for the as-yet unborn branch.

As part of the implicit initialization of the local repository, Git
just learned to respect `init.defaultBranch` to choose a different main
branch name. We now really want that branch name to be used as a
fall-back.

At the same time, we also want to make sure that `core.mainBranch` is
set, correctly reflecting the name of the main branch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>