TODO clone: handle overridden main branch names
TODO: make sure that `core.mainBranch` is recorded if the remote has
one.

TODO: teach `git branch -m` to set `branch.mainBranch` if we're on the
main branch.

When cloning a repository without any branches, Git chooses a default
branch name for the as-yet unborn branch.

As part of the implicit initialization of the local repository, Git
just learned to respect `init.defaultBranch` to choose a different main
branch name. We now really want that branch name to be used as a
fall-back.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>